requirement.txt
adjustText==1.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
altair @ file:///C:/b/abs_0b1avjyzoc/croot/altair_1743016738596/work
annotated-types==0.7.0
anyio==4.9.0
attrs @ file:///C:/b/abs_89hmquz5ga/croot/attrs_1734533130810/work
backoff==2.2.1
blinker @ file:///C:/b/abs_b1i87khtob/croot/blinker_1737448732095/work
blis==1.3.0
Bottleneck @ file:///C:/b/abs_816hr2khp1/croot/bottleneck_1731058648110/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
cachetools @ file:///C:/b/abs_fcvwpc5vh9/croot/cachetools_1738224687938/work
catalogue==2.0.10
certifi @ file:///C:/b/abs_8a944p1_gn/croot/certifi_1738623753421/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///C:/b/abs_42rbjngfys/croot/click_1744271606983/work
cloudpathlib==0.21.0
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1699472650914/work
confection==0.1.5
contourpy==1.3.2
country_converter==1.3
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.2
fsspec==2025.3.2
gitdb @ file:///tmp/build/80754af9/gitdb_1617117951232/work
GitPython @ file:///C:/b/abs_2bkslnqz4i/croot/gitpython_1720455044865/work
greenlet==3.2.3
grpcio==1.72.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.2
hyperframe==6.1.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema @ file:///C:/b/abs_394_t6__xq/croot/jsonschema_1728486718320/work
jsonschema-specifications @ file:///C:/Users/dev-admin/py312/jsonschema-specifications_1706803038066/work
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-openai==0.3.21
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.45
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
marshmallow==3.26.1
matplotlib==3.10.1
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath==1.3.0
multidict==6.4.4
murmurhash==1.0.12
mypy_extensions==1.1.0
narwhals @ file:///C:/b/abs_224ub5a528/croot/narwhals_1742845975749/work
networkx==3.4.2
numexpr @ file:///C:/b/abs_05o8p7bfml/croot/numexpr_1730215959182/work
numpy @ file:///C:/b/abs_71ejae7e2v/croot/numpy_and_numpy_base_1742570444043/work/dist/numpy-2.2.4-cp312-cp312-win_amd64.whl#sha256=51cd6ca08057acb115487cc25926841a5f12f8555a06e60179ffc632b9c70638
openai==1.75.0
orjson==3.10.18
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas @ file:///C:/b/abs_e18w9ay4we/croot/pandas_1732735115740/work/dist/pandas-2.2.3-cp312-cp312-win_amd64.whl#sha256=2c8058e41dda2e7ecfff0740ad1d98f17d528c740f71b038c523a9d24d2321b0
pillow @ file:///C:/b/abs_68t826txdy/croot/pillow_1744613085333/work
portalocker==2.10.1
preshed==3.0.9
propcache==0.3.1
protobuf @ file:///C:/b/abs_b5v84rwh56/croot/protobuf_1742419721750/work/bazel-bin/python/dist/protobuf-5.29.3-cp312-abi3-win_amd64.whl#sha256=ea7ea1d0989645071f9db2e1e8cdbb05a8f5fe001144842c6bff5b642d4d394e
pyarrow @ file:///C:/b/abs_0aneze9pgl/croot/pyarrow_1738229440651/work/python
pycountry==24.6.1
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck @ file:///C:/b/abs_47xw5z_59x/croot/pydeck_1738340983078/work
Pygments==2.19.1
pymongo==4.12.0
pyparsing==3.2.3
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
python-dotenv==1.1.0
python-multipart==0.0.20
pytz @ file:///C:/b/abs_6ap4tsz1ox/croot/pytz_1713974360290/work
pywin32==310
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
qdrant-client==1.14.2
referencing @ file:///C:/Users/dev-admin/py312/referencing_1706802962559/work
regex==2024.11.6
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.3
rpds-py @ file:///C:/b/abs_0c6z5kcdb6/croot/rpds-py_1736545465023/work
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==72.1.0
shellingham==1.5.4
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
smart-open==7.1.0
smmap @ file:///tmp/build/80754af9/smmap_1611694433573/work
sniffio==1.3.1
spacy==3.8.5
spacy-alignments==0.9.1
spacy-huggingface-hub==0.0.10
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.8
SQLAlchemy==2.0.41
srsly==2.5.1
starlette==0.46.2
streamlit @ file:///C:/b/abs_844t7d2z7u/croot/streamlit-split_1743683308272/work/streamlit
sympy==1.14.0
tenacity @ file:///C:/b/abs_1749lsqys6/croot/tenacity_1730304400080/work
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==2.7.0
tornado @ file:///C:/b/abs_7cyu943ybx/croot/tornado_1733960510898/work
tqdm==4.67.1
transformers==4.49.0
typer==0.7.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
uvicorn==0.34.2
wasabi==1.1.3
watchdog @ file:///C:/b/abs_29_7vso69r/croot/watchdog_1740409940307/work
watchfiles==1.0.5
weasel==0.4.1
websockets==15.0.1
wheel==0.45.1
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
wrapt==1.17.2
yarl==1.20.0
zstandard==0.23.0
